+++ packages/moonstone/CHANGELOG.md
@@ -11,6 +11,13 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Fixed
 
 - `moonstone/IncrementSlider` so that the knob is spottable via pointer, and 5-way navigation between the knob and the increment/decrement buttons is functional
+### Changed
+
+- `moonstone/ProgressBar` properties `progress` and `backgroundProgress` to accept a number between 0 and 1.
+
+### Removed
+
+- `moonstone/ProgressBar` properties `min` and `max`
 
 ## [1.0.0-beta.1] - 2016-12-30